refactor(#4712): pure-move de-giant discord/formatting.rs
